requirements.txt
discord.py==1.3.2
request==2.23.0